requirements.txt
simplejson
psycopg2-binary
pandas
sqlalchemy